warnings fixed, added polylines,texts,ellipse support

+++ gdal/ogr/ogrsf_frmts/cad/ogr_cad.h
@@ -63,7 +63,7 @@ class OGRCADDataSource : public GDALDataset
     std::unique_ptr<CADFile>    spoCADFile;
     
     OGRCADLayer                 **papoLayers;
-    size_t                      nLayers;
+    int                         nLayers;
     
 public:
     OGRCADDataSource();
